inspector: Move JS API code to separate file
